Update installation instructions for newer cabal-install versions. Refs #6.
Installation with newer versions of cabal-install no longer uses
sandboxes.

This commit replaces the installation instructions for a set that is
working with cabal-install 3.6.

The installation requires installing BNFC separately first, since adding
it to build-tool-depends or just to the same installation command as
ogma results in compilation errors. In particular, when BNFC is
installed together with ogma in one command:

$ cabal install BNFC ogma-cli

the result is that cabal won't symlink the BNFC executable, so
ogma-language-* packages cannot find it.